+++ src/Order/Checkout/CheckoutPage.php
@@ -1,11 +1,13 @@
 <?php
 declare(strict_types=1);
 
-namespace SwipeStripe\Pages;
+namespace SwipeStripe\Order\Checkout;
+
+use SwipeStripe\Pages\RequiredSinglePage;
 
 /**
  * Class CheckoutPage
- * @package SwipeStripe\Pages
+ * @package SwipeStripe\Order\Checkout
  */
 class CheckoutPage extends \Page
 {
