+++ templates/SwipeStripe/Order/Checkout/Layout/CheckoutPage.ss
